add test for an empty string and rescue code
IRB has a subtle difference in behavior when it comes to encodings. An Encoding::ConverterNotFoundError is not raised during forced conversions to UTF-7 and ISO-2022-JP-2. On Ruby 2.0 and Ruby 1.9,3(MRI, OSX10.9) these conversions raise an exception and prevent the test suite from completing its run when evaluated in the default US-ASCII format. IRB is natively encoded in the UTF-8 format on a MAC. A simple trick to encapsulate the preferred IRB Behavior is to open a pipe to an external shell and compare its output for '\x2E'. This behavior is also repeated and tested in the 2nd test